% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{approvals}
\alias{approvals}
\title{Retrieve 'approvals' Data as a Tibble}
\usage{
approvals()
}
\value{
A tibble containing the 'approvals' data.
}
\description{
This function is a wrapper for `load_data()` that specifically retrieves the 'approvals' data.
It loads the 'approvals' data object and coerces it into a tibble before returning.
}
\examples{
\dontrun{
approvals_data <- approvals()
}
}
