% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{confirm_slash}
\alias{confirm_slash}
\title{Ensure Trailing Slash}
\usage{
confirm_slash(path)
}
\arguments{
\item{path}{A character string representing a file path.}
}
\value{
A character string of the file path, guaranteed to end with a forward slash.
}
\description{
`confirm_slash` ensures that a given file path ends with a forward slash.
This is useful for constructing file paths in a consistent manner, especially
when working across different operating systems. If the path already ends with
a forward slash, it remains unchanged.
}
\examples{
confirm_slash("path/to/directory")
confirm_slash("path/to/directory/")

}
