% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_path}
\alias{data_path}
\title{Set the Default Data Path for databased}
\usage{
data_path(path)
}
\arguments{
\item{path}{A character string specifying the file path to be set as the default data path.}
}
\description{
This function sets the default data path for the `databased` package by updating the
`databased.path` option. This path is used as the default location to read from and write to data files.
}
\details{
`data_path` is a utility function that modifies the global option `databased.path`.
This option is used by the `databased` package to determine the default directory
for data operations. Setting this path is especially useful for ensuring
consistency in file locations across different sessions and scripts.

It's important to note that this function changes a global option, and the set path
remains in effect for the duration of the R session or until it's modified again.
}
\examples{
\dontrun{
data_path("path/to/your/data")
}

}
