% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{database_it}
\alias{database_it}
\title{Save an R object to a specified database path}
\usage{
database_it(obj, envir = .GlobalEnv)
}
\arguments{
\item{obj}{The object to be saved.}

\item{envir}{The environment from which the object should be saved (default is the global environment).}
}
\value{
Invisible NULL; the function is called for its side effect of saving a file.
}
\description{
This function saves an R object to a predefined path in an '.rda' format. The object
is saved using its name in the environment specified, allowing for easy retrieval using
the same name. The default save path is set by the global 'path_database' variable.
}
\examples{
\dontrun{
  my_data <- mtcars
  database_it(my_data)
  # This will save the 'my_data' object to 'S:/data/databased/rda/my_data.rda'
}
}
