% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{file_tibble_log}
\alias{file_tibble_log}
\title{Generate a Tibble of Log Files}
\usage{
file_tibble_log(
  path = "",
  file_type = ".",
  recursive = FALSE,
  filter_out_tilda = TRUE
)
}
\arguments{
\item{path}{A string indicating a sub-path within `x` to further narrow down the file search.}

\item{file_type}{A string indicating the type of the files to include represented by their extension.}

\item{recursive}{A logical indicating whether to perform a recursive search in subdirectories.}

\item{filter_out_tilda}{A logical indicating whether to filter out files starting with a tilde (~), which are often temporary or backup files.}
}
\value{
A tibble with each log file and its associated metadata.
}
\description{
This function is a wrapper for `file_tibble_x` preset to search for log files.
}
\examples{
\dontrun{
file_tibble_log()
}
}
