% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{file_tibble_x}
\alias{file_tibble_x}
\title{Generate a Tibble of File Information}
\usage{
file_tibble_x(
  x,
  path = "",
  file_type = ".",
  recursive = FALSE,
  filter_out_tilda = TRUE
)
}
\arguments{
\item{x}{A string specifying the subdirectory within the base path to search in.}

\item{path}{A string indicating a sub-path within `x` to further narrow down the file search.}

\item{file_type}{A string indicating the type of the files to include represented by their extension.}

\item{recursive}{A logical indicating whether to perform a recursive search in subdirectories.}

\item{filter_out_tilda}{A logical indicating whether to filter out files starting with a tilde (~), which are often temporary or backup files.}
}
\value{
A tibble with each file and its associated metadata.
}
\description{
This internal function generates a data frame (tibble) that contains the name of each file in the specified directory,
along with the path and some metadata, filtered by a particular subdirectory and file type.
}
\examples{
\dontrun{
file_tibble_x("subdir")
}
}
