% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_log}
\alias{get_log}
\title{Load a Log File}
\usage{
get_log(log_name)
}
\arguments{
\item{log_name}{The name of the log file to be loaded (without the `.rda` extension).}
}
\value{
The object stored in the log file, typically a data frame or a list.
        This object should have the same name as `log_name`.
}
\description{
`get_log` loads a log file stored in R data format (`.rda`). It constructs the file path
based on the provided log name and a base path set as an option. The function then
loads the log data into the R environment. It is assumed that the log file contains
an object with the same name as `log_name`.
}
\examples{
\dontrun{
  get_log("example_log") # Loads 'example_log.rda' from the log directory
}

}
