% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load_data}
\alias{load_data}
\title{Load Data from a Specified Path}
\usage{
load_data(data_name, extension = ".rda")
}
\arguments{
\item{data_name}{The name of the dataset to load.}

\item{extension}{The file extension of the dataset, defaulting to '.rda'.}
}
\value{
The specified data object.
}
\description{
This function loads an R data file (.rda) from a predefined path into the global environment.
The path is combined with the data name and extension to locate the file.
If the file or the specified data object within the file does not exist, the function stops with an error.
}
\examples{
\dontrun{
load_data("approvals")
}
}
