% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{log_it}
\alias{log_it}
\title{Append Entry to Log and Save}
\usage{
log_it(log_entry, log_name)
}
\arguments{
\item{log_entry}{A data frame or similar structure representing the new log entry to be appended.}

\item{log_name}{The name of the log file (without the `.rda` extension) to which the log entry will be appended.}
}
\value{
The function does not return a value but updates the log file specified by `log_name`.
}
\description{
The `log_it` function appends a new entry to an existing log and then saves
the updated log back to its file. It relies on the `get_log` function to
retrieve the current log and uses tidyverse functions for manipulation.
}
\examples{
\dontrun{
  new_entry <- tibble(timestamp = Sys.time(), message = "New log entry")
  log_it(new_entry, "example_log")
}

}
