% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{options_to_clipboard}
\alias{options_to_clipboard}
\title{Copy Database Path Setting to Clipboard}
\usage{
options_to_clipboard()
}
\value{
The function does not return a value but performs an action: copying text to the clipboard.
}
\description{
This function copies a specific R options setting command to the clipboard.
The command sets the `databased.path` option to a predefined path (`"S:/data/databased"`).
This can be useful for quickly sharing or applying this setting in different R scripts
or R sessions.
}
\details{
The function utilizes the `clipr` package to interact with the system clipboard.
It is designed to work in environments where clipboard access is available.
The specific text copied to the clipboard is:
`options(databased.path = "S:/data/databased")`.
}
\examples{
\dontrun{
  options_to_clipboard() # Copies the options command to the clipboard
}

}
