% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{projects}
\alias{projects}
\title{Retrieve 'projects' Data as a Tibble}
\usage{
projects()
}
\value{
A tibble containing the 'projects' data.
}
\description{
This function is a wrapper for `load_data()` that specifically retrieves the 'projects' data.
It loads the 'projects' data object and coerces it into a tibble before returning.
}
\examples{
\dontrun{
projects_data <- projects()
}
}
