% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_read_in.R
\name{read_excel_all_internal}
\alias{read_excel_all_internal}
\title{Read All Sheets from an Excel File}
\usage{
read_excel_all_internal(path, sheets = NULL, sheets_regex = ".")
}
\arguments{
\item{path}{A character string specifying the path to the Excel file.}

\item{sheets}{A vector with sheet names or positions (numeric). If NULL (the default), the entire workbook is read.}

\item{sheets_regex}{A regex string used to select the sheets desired.}
}
\value{
A named list of data frames. Each data frame corresponds to a sheet in the Excel file.
}
\description{
This function reads all sheets from a given Excel file and returns a named list of data frames
where each sheet corresponds to an entry in the list.
}
\examples{
\dontrun{
  all_sheets <- read_excel_all_internal("path_to_file.xlsx")
  head(all_sheets$Sheet1)
}

}
\keyword{internal}
