% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{set_databased_path}
\alias{set_databased_path}
\title{Set databased.path in .Rprofile}
\usage{
set_databased_path(path)
}
\arguments{
\item{path}{A character string specifying the desired path for the `databased.path` option.}
}
\description{
Modifies the `.Rprofile` file in the user's home directory to set the `databased.path` option.
This function ensures that `databased.path` is correctly set for future R sessions.
}
\details{
- This function alters the `.Rprofile` file located in the user's home directory.
- If the `.Rprofile` file does not exist, it will be created with the necessary setting.
- If the file exists, the function will append or modify the `databased.path` setting
  while preserving other settings.
- The `data_path()` function can be used as an alternative to set the `databased.path`
  option for the current session without altering the `.Rprofile` file.
- Note that a `.Rprofile` file in the project directory will take precedence over the
  settings made by this function.

IMPORTANT
- Modifying the `.Rprofile` can impact the behavior of all R sessions for the user.
  Ensure that you understand the implications of these changes.
- Consider backing up your `.Rprofile` before making automated modifications.
}
\examples{
\dontrun{
# Set the databased.path in .Rprofile
set_databased_path("path/to/databased")
}

}
