% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{slice_changes}
\alias{slice_changes}
\title{Detect and Retain Rows with Changes in a DataFrame}
\usage{
slice_changes(df, unique_identifier, ignore_cols)
}
\arguments{
\item{df}{A dataframe containing the data to be checked for changes.}

\item{unique_identifier}{The name of the column that uniquely identifies each row or group for comparison. This parameter is treated as a grouping variable.}

\item{ignore_cols}{A vector of column names to be ignored during change detection.}
}
\value{
A dataframe containing only the rows where changes have been detected in any of the columns, excluding those specified in `ignore_cols`.
}
\description{
This function identifies and retains rows within a dataframe where changes have occurred in any column, excluding specified columns. It's designed to work within groups defined by a unique identifier and does not consider changes from or to `NA` as changes.
}
\examples{
\dontrun{
# Assuming 'data' is a dataframe with a unique ID column 'id' and you
# want to ignore 'timestamp' column.
changes_df <- slice_changes(data, unique_identifier = "id", ignore_cols = "timestamp")

# The above will return a dataframe with rows where changes occurred.
}
}
