% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulation.R
\name{snapshot_data}
\alias{snapshot_data}
\title{Take a Snapshot of Data as of a Specific Date}
\usage{
snapshot_data(
  df,
  as_of = NULL,
  unique_identifier = project_id,
  date_data = date_data
)
}
\arguments{
\item{df}{A dataframe to take a snapshot from.}

\item{as_of}{A date in as_date format or date character string in 'yyyy-mm-dd' format (which is converted with as_date) specifying the snapshot date. If `NULL`, the function will use the latest date available in `date_data`.}

\item{unique_identifier}{The column name that uniquely identifies each entry or group in `df`. If not provided, defaults to 'project_id'.}

\item{date_data}{The column name containing date information used to filter the data.}
}
\value{
A dataframe that is a snapshot of `df` as of `as_of`, with NA values filled down and only the most recent data for each unique identifier.
}
\description{
This function filters a dataframe to represent a 'snapshot' of the data as of a particular date. It fills in NA values downwards and retains only the most recent entry for each unique identifier up to the specified date.
}
\examples{
\dontrun{
# Assuming 'data_frame' is a dataframe with 'project_id' as the unique
# identifier and 'record_date' as the date column:
snapshot <- snapshot_data(data_frame, as_of = "2021-12-31",
unique_identifier = "project_id",
date_data = "record_date")

# This will return a snapshot of 'data_frame' as of 2021-12-31, with all
# NA values filled down and only the latest record for each project_id.
}
}
